Exec dcrwallet to perform wallet creation
dcrwallet's loader and prompt packages are part of the main module and
are not intended, recommended, or versioned for external consumption.
Exec the extracted dcrwallet executable instead with the correct
arguments for the dcrinstall settings.  This avoids problems due to
dcrinstall performing a different wallet creation process than what
dcrwallet actually does.