Only self-check version in release builds
This commit disables the comparison of dcrinstall's own embedded
version against the version discovered in the 'latest' file unless the
version is embedded using a linker flag by the release builder.
Normal development builds of dcrinstall will no longer self-check, and
the release builder should never produce a dcrinstall executable which
is incompatible with the manifests it creates.

While here, also update the fallback manifest URLs for the decred and
dexc distributions for the v1.6.0-rc3 release.