Self-check version using filename
It should not matter where the file is hosted to determine if an
upgrade is necessary, but only the filename of the dcrinstall
manifest.  This is because the filename contains the release version.

Relaxing this rule allows for better testing, as faked 'latest' files
with file:// urls won't trip the version self check.