always install dex, never download/install bitcoin
This updates dcrinstall so that it always installs dex, deprecating
the --dcrdex flag.

This also removes all bitcoin download and install actions, which
aren't needed or recommended.

Update CI to use latest Go versions and linter.

Add the --allowrunning switch to not fail if a running process with
the name of one of the binaries is found. There's nothing in the
ps output to suggest it's the same process, and it doesn't really
make a difference if it is.